requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.0
langchain>=0.1.10
langchain-community>=0.0.25
langchain-openai>=0.0.8
langchain-core>=0.1.28
langchain-text-splitters>=0.0.1
chromadb>=0.4.24
openai>=1.13.0
tiktoken>=0.6.0
pypdf>=4.0.0
httpx>=0.27.0
python-multipart>=0.0.9
python-multipart>=0.0.9
scikit-learn>=1.4.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
redis>=5.0.0

